Break out some logic to new service files to clean things up a little.
